XGPaint

XGPaint.AbstractForegroundModel

XGPaint.CIB_Planck2013

XGPaint.Enmap

XGPaint.Radio_Sehgal2009

XGPaint.build_c_lnm2r_interpolator

XGPaint.build_r2z_interpolator

XGPaint.build_shang_interpolator

XGPaint.build_sigma_sat_ln_interpolator

XGPaint.chunk

XGPaint.ellpad

XGPaint.enfft!

XGPaint.enifft!

XGPaint.generate_sources

XGPaint.generate_sources

XGPaint.generate_subhalo_offsets

XGPaint.get_angles

XGPaint.get_basic_halo_properties

XGPaint.get_cosmology

XGPaint.get_interpolators

XGPaint.hod_sehgal

XGPaint.integrand_L

XGPaint.l2f

XGPaint.m2r

XGPaint.mz2c

XGPaint.paint!

XGPaint.paint!

XGPaint.process_centrals!

XGPaint.process_sats!

XGPaint.read_halo_catalog_hdf5

XGPaint.sehgal_LF!

XGPaint.shang_z_evo

XGPaint.AbstractForegroundModel — Type

All foreground models inherit from this type.

XGPaint.CIB_Planck2013 — Type

CIB_Planck2013{T}(model parameters...)

Define CIB model parameters. Defaults are from Viero et al. 2013.

model = CIB_Planck2013{Float32}(shang_Mpeak=10^12.4)

XGPaint.Enmap — Type

Map type, contains an AbstractArray and a WCS object, but behaves like the AbstractArray it contains for array operations.

It only implements the subset of Base.Array operations which are common on maps. You should work with the data directly using enmap_instance.data if you need additional Array functions.

XGPaint.Radio_Sehgal2009 — Type

Radio_Sehgal2009{T}(model parameters...)

Define CIB model parameters. Defaults are from Viero et al. 2013.

model = CIBModel{Float32}(a_0=0.4)

XGPaint.build_c_lnm2r_interpolator — Method

Generates an interpolator r(c, lnm)

Generate a LinearInterpolation object that turns concentration and ln(M_halo) into satellite radius.

XGPaint.build_r2z_interpolator — Method

Construct a fast r2z linear interpolator.

XGPaint.build_shang_interpolator — Method

Build a linear interpolation function which maps log(Mh) to Nsat.

XGPaint.build_sigma_sat_ln_interpolator — Method

Build a linear interpolator that takes in ln(M_halo) and returns sigma.

XGPaint.chunk — Method

Generates a list of tuples which describe starting and ending chunk indices. Useful for parallelizing an array operation.

XGPaint.ellpad — Method

Utility function which prepends some zeros to an array. It makes a copy instead of modifying the input.

XGPaint.enfft! — Method

Physically normalized enmap FFT

XGPaint.enifft! — Method

Physically normalized enmap inverse FFT

XGPaint.generate_sources — Method

Produce a source catalog from a model and halo catalog.

XGPaint.generate_sources — Method

Produce a source catalog from a model and halo catalog.

XGPaint.generate_subhalo_offsets — Method

Generate an array where the value at index i corresponds to the index of the first source of halo i. Takes an array where the value at index i corresponds to the number of subhalos that halo i has.

XGPaint.get_angles — Method

Compute angles of halos

XGPaint.get_basic_halo_properties — Method

Fill in basic halo properties.

XGPaint.get_cosmology — Method

Construct a background cosmology.

This function duplicates the cosmology() function in Cosmology.jl, but with typing. The type of the cosmology will the type of h and OmegaM. This is primarily for keeping the code entirely in Float32 or Float64.

XGPaint.get_interpolators — Method

Construct the necessary interpolator set.

XGPaint.hod_sehgal — Method

Populate halos with radio sources according to the HOD in Sehgal et al. 2009.

The optional rng parameter provides an array of random number generators, one for each thread.

XGPaint.integrand_L — Method

<L_sat> interpolation values

XGPaint.l2f — Method

Inverse square law with redshift dependence.

XGPaint.m2r — Method

Convert virial mass to virial radius.

XGPaint.mz2c — Method

Compute concentration factor from Duffy et al. 2008.

XGPaint.paint! — Method

Paint a source catalog onto a map.

This function creates the arrays for you.

XGPaint.paint! — Method

Paint a source catalog onto a map, recording the fluxes.

XGPaint.process_centrals! — Method

Fill up arrays with information related to CIB central sources.

XGPaint.process_sats! — Method

Fill up arrays with information related to CIB satellites.

XGPaint.read_halo_catalog_hdf5 — Method

Utility function to read an HDF5 table with x, y, z, M_h as the four rows. The hdf5 record is "halos".

XGPaint.sehgal_LF! — Method

Fills the result array with draws from the luminosity function.

XGPaint.shang_z_evo — Method

Compute redshift evolution factor for LF.