Home

XGPaint.jl

XGPaint.AbstractForegroundModel

XGPaint.CIBModel

XGPaint.build_c_lnm2r_interpolator

XGPaint.build_r2z_interpolator

XGPaint.build_shang_interpolator

XGPaint.build_sigma_sat_ln_interpolator

XGPaint.get_cosmology

XGPaint.get_interpolators

XGPaint.integrand_L

XGPaint.l2f

XGPaint.m2r

XGPaint.mz2c

XGPaint.read_halo_catalog_hdf5

XGPaint.shang_z_evo

XGPaint.CIBModel — Type.

CIBModel{T}(model parameters...)

Define CIB model parameters. Defaults are from Viero et al. 2013.

model = CIBModel{Float32}(shang_Mpeak=10^12.4)

source

XGPaint.build_c_lnm2r_interpolator — Method.

Generates an interpolator r(c, lnm)

Generate a LinearInterpolation object that turns concentration and ln(M_halo) into satellite radius.

source

XGPaint.build_r2z_interpolator — Method.

Construct a fast r2z linear interpolator.

source

XGPaint.build_shang_interpolator — Method.

Build a linear interpolation function which maps log(Mh) to Nsat.

source

XGPaint.build_sigma_sat_ln_interpolator — Method.

Build a linear interpolator that takes in ln(M_halo) and returns sigma.

source

XGPaint.get_cosmology — Method.

Construct a background cosmology.

This function duplicates the cosmology() function in Cosmology.jl, but with typing. The type of the cosmology will the type of h and OmegaM. This is primarily for keeping the code entirely in Float32 or Float64.

source

XGPaint.get_interpolators — Method.

Construct the necessary interpolator set.

source

XGPaint.l2f — Method.

Inverse square law with redshift dependence.

source

XGPaint.m2r — Method.

Convert virial mass to virial radius.

source

XGPaint.mz2c — Method.

Compute concentration factor from Duffy et al. 2008.

source

XGPaint.read_halo_catalog_hdf5 — Method.

Utility function to read an HDF5 table with x, y, z, M_h as the four rows. The hdf5 record is "halos".

source

XGPaint.AbstractForegroundModel — Type.

All foreground models inherit from this type.

source

XGPaint.integrand_L — Method.

<L_sat> interpolation values

source

XGPaint.shang_z_evo — Method.

Compute redshift evolution factor for LF.

source