Home

XGPaint.jl

XGPaint.AbstractForegroundModel

XGPaint.CIB_Planck2013

XGPaint.Radio_Sehgal2009

XGPaint.build_c_lnm2r_interpolator

XGPaint.build_r2z_interpolator

XGPaint.build_shang_interpolator

XGPaint.build_sigma_sat_ln_interpolator

XGPaint.chunk_list

XGPaint.generate_sources

XGPaint.generate_sources

XGPaint.get_cosmology

XGPaint.get_interpolators

XGPaint.get_thread_RNG

XGPaint.hod_sehgal!

XGPaint.integrand_L

XGPaint.l2f

XGPaint.m2r

XGPaint.mz2c

XGPaint.paint!

XGPaint.process_centrals!

XGPaint.process_sats!

XGPaint.read_halo_catalog_hdf5

XGPaint.sehgal_LF!

XGPaint.shang_z_evo

XGPaint.CIB_Planck2013 — Type.

CIB_Planck2013{T}(model parameters...)

Define CIB model parameters. Defaults are from Viero et al. 2013.

model = CIB_Planck2013{Float32}(shang_Mpeak=10^12.4)

source

XGPaint.Radio_Sehgal2009 — Type.

Radio_Sehgal2009{T}(model parameters...)

Define CIB model parameters. Defaults are from Viero et al. 2013.

model = CIBModel{Float32}(a_0=0.4)

source

XGPaint.build_c_lnm2r_interpolator — Method.

Generates an interpolator r(c, lnm)

Generate a LinearInterpolation object that turns concentration and ln(M_halo) into satellite radius.

source

XGPaint.build_r2z_interpolator — Method.

Construct a fast r2z linear interpolator.

source

XGPaint.build_shang_interpolator — Method.

Build a linear interpolation function which maps log(Mh) to Nsat.

source

XGPaint.build_sigma_sat_ln_interpolator — Method.

Build a linear interpolator that takes in ln(M_halo) and returns sigma.

source

XGPaint.chunk_list — Method.

Generates a list of tuples which describe starting and ending chunk indices. Useful for parallelizing an array operation.

source

XGPaint.generate_sources — Method.

Produce a source catalog from a model and halo catalog.

source

XGPaint.generate_sources — Method.

Produce a source catalog from a model and halo catalog.

source

XGPaint.get_cosmology — Method.

Construct a background cosmology.

This function duplicates the cosmology() function in Cosmology.jl, but with typing. The type of the cosmology will the type of h and OmegaM. This is primarily for keeping the code entirely in Float32 or Float64.

source

XGPaint.get_interpolators — Method.

Construct the necessary interpolator set.

source

XGPaint.l2f — Method.

Inverse square law with redshift dependence.

source

XGPaint.m2r — Method.

Convert virial mass to virial radius.

source

XGPaint.mz2c — Method.

Compute concentration factor from Duffy et al. 2008.

source

XGPaint.paint! — Method.

Paint a source catalog onto a map.

source

XGPaint.read_halo_catalog_hdf5 — Method.

Utility function to read an HDF5 table with x, y, z, M_h as the four rows. The hdf5 record is "halos".

source

XGPaint.AbstractForegroundModel — Type.

All foreground models inherit from this type.

source

XGPaint.get_thread_RNG — Method.

Generate an array of random number generators, for each thread.

source

XGPaint.hod_sehgal! — Method.

Populate halos with radio sources according to the HOD in Sehgal et al. 2009.

The optional rng parameter provides an array of random number generators, one for each thread. If none are specified, this function will call XGPaint.get_thread_RNG() to create such an array, which takes about 1 second.

source

XGPaint.integrand_L — Method.

<L_sat> interpolation values

source

XGPaint.process_centrals! — Method.

Fill up arrays with information related to CIB central sources.

source

XGPaint.process_sats! — Method.

Fill up arrays with information related to CIB satellites.

source

XGPaint.sehgal_LF! — Method.

Fills the result array with draws from the luminosity function.

source

XGPaint.shang_z_evo — Method.

Compute redshift evolution factor for LF.

source