XGPaint

XGPaint paints maps of extragalactic foregrounds using halo models. We provide CIB and radio models. The CIB model is described in the Websky paper, Stein et al. 2020. The radio model is from Li et al. 2021, a paper in prep for the Websky suite, and is derived from Sehgal et al. 2009.

The general workflow is:

Read halos

Specify a background cosmology and source model

Use the model to put sources in halos

Put the sources on maps

General Methods

There are a few functions that are generally applicable for all foreground models.

XGPaint.read_halo_catalog_hdf5 — Function

Utility function to read an HDF5 table with x, y, z, M_h as the four rows. The hdf5 record is "halos".

XGPaint.get_cosmology — Function

get_cosmology(::Type{T}; h=0.69, Neff=3.04, OmegaK=0.0,
    OmegaM=0.29, OmegaR=nothing, Tcmb=2.7255, w0=-1, wa=0)

Construct a background cosmology. This function duplicates the cosmology() function in Cosmology.jl, but with the numeric type specified.

Arguments:

::Type{T}: numerical type to use for calculations

Keywords

h - Dimensionless Hubble constant

OmegaK - Curvature density (Ω_k)

OmegaM - Matter density (Ω_m)

OmegaR - Radiation density (Ω_r)

Tcmb - CMB temperature in Kelvin; used to compute Ω_γ

Neff - Effective number of massless neutrino species; used to compute Ω_ν

w0 - CPL dark energy equation of state; w = w0 + wa(1-a)

wa - CPL dark energy equation of state; w = w0 + wa(1-a)

Example

julia> get_cosmology(Float32; h=0.7)
Cosmology.FlatLCDM{Float32}(0.7f0, 0.7099147f0, 0.29f0, 8.5307016f-5)